Return Sheets logic, separate verify_status
- Add required keywords for global variables
- Reorganize into standard and business custom modules
- Keep config as the cross-file mutable variable storage location
